CRUD complet et authorization implemente

+++ API/BLL/BLLPassword.cs
@@ -1,5 +1,6 @@
 ﻿using API.Cryptography;
 using API.Database;
+using API.Helper;
 using API.Models.V1;
 
 namespace API.BLL
@@ -17,15 +18,17 @@ namespace API.BLL
 
   public class BLLPassword : IBLLPassword
   {
-    private readonly IPasswordRepository _passwordRepository;
     private readonly IUserRepository _userRepository;
+    private readonly IPasswordRepository _passwordRepository;
     private readonly IVaultRepository _vaultRepository;
+    private readonly IAuthorizationHelper _authorizationHelper;
 
-    public BLLPassword(IPasswordRepository passwordRepository, IVaultRepository vaultRepository, IUserRepository userRepository)
+    public BLLPassword(IPasswordRepository passwordRepository, IVaultRepository vaultRepository, IUserRepository userRepository, IAuthorizationHelper authorizationHelper)
     {
       _passwordRepository = passwordRepository;
       _vaultRepository = vaultRepository;
       _userRepository = userRepository;
+      _authorizationHelper = authorizationHelper;
     }
 
     public async Task<int> CreatePassword(IPasswordViewModel passwordViewModelToCreate)
@@ -35,16 +38,22 @@ namespace API.BLL
         throw new Exception("Application_name is required.");
       }
 
-      if (await _vaultRepository.GetVaultById(passwordViewModelToCreate.Vault_id) == null)
+      IVaultModel vaultInDatabase = await _vaultRepository.GetVaultById(passwordViewModelToCreate.Vault_id);
+      if (vaultInDatabase == null)
       {
         throw new Exception($"The vault with id:{passwordViewModelToCreate.Vault_id} doesn't exist!");
+      }
+
+      if (await _authorizationHelper.IsRefusedToPerformActionOnUser(vaultInDatabase.User_id))
+      {
+        throw new Exception($"You don't have the authorization!");
       }
 
       IPasswordModel passwordModelToCreate = _passwordRepository.GeneratePasswordModelFromPasswordViewModel(passwordViewModelToCreate);
       int createdPasswordId = await _passwordRepository.CreatePassword(passwordModelToCreate);
 
-      bool createdRelationShip = await _passwordRepository.CreateRelationshipMember(passwordViewModelToCreate.Vault_id, createdPasswordId);
-      if (!createdRelationShip)
+      bool isCreatedRelationShipFailed = !await _passwordRepository.CreateRelationshipMember(passwordViewModelToCreate.Vault_id, createdPasswordId);
+      if (isCreatedRelationShipFailed)
       {
         throw new Exception($"Couln't create the relationship between vault:{passwordViewModelToCreate.Vault_id} and password:{createdPasswordId} !");
       }
@@ -54,6 +63,23 @@ namespace API.BLL
 
     public async Task<bool> DeletePasswordById(int password_id)
     {
+      IPasswordModel passwordInDatabase = await _passwordRepository.GetPasswordById(password_id);
+      if (passwordInDatabase == null)
+      {
+        throw new Exception($"The password with the id: {password_id} doesn't exist!");
+      }
+
+      IVaultModel vaultInDatabase = await _vaultRepository.GetVaultById(passwordInDatabase.Vault_id);
+      if (vaultInDatabase == null)
+      {
+        throw new Exception($"The vault doesn't exist!");
+      }
+
+      if (await _authorizationHelper.IsRefusedToPerformActionOnUser(vaultInDatabase.User_id))
+      {
+        throw new Exception($"You don't have the authorization!");
+      }
+
       return await _passwordRepository.DeletePasswordById(password_id);
     }
 
@@ -71,6 +97,17 @@ namespace API.BLL
         return false;
       }
 
+      IVaultModel vaultInDatabase = await _vaultRepository.GetVaultById(passwordInDatabase.Vault_id);
+      if (vaultInDatabase == null)
+      {
+        throw new Exception($"The vault doesn't exist!");
+      }
+
+      if (await _authorizationHelper.IsRefusedToPerformActionOnUser(vaultInDatabase.User_id))
+      {
+        throw new Exception($"You don't have the authorization!");
+      }
+
       if (passwordInDatabase.Application_name != passwordToPatch.Application_name)
       {
         passwordInDatabase.Application_name = passwordToPatch.Application_name;
@@ -81,7 +118,7 @@ namespace API.BLL
         passwordInDatabase.Username = passwordToPatch.Username;
       }
 
-      if (!string.IsNullOrWhiteSpace(passwordToPatch.Password))
+      if (string.IsNullOrWhiteSpace(passwordToPatch.Password))
       {
         passwordInDatabase.Encrypted_password = CryptographyUtil.Instance.EncryptasswordForString(passwordToPatch.Password, Enum.EnumPasswordComplexity.Medium);
       }
